requirements.txt
pandas-datareader==0.10.0
pandas-flavor==0.8.1
scikit-learn==1.8.0
yfinance==1.0
plotly==6.5.2
kaleido==0.2.1

# attrs==25.4.0
# beautifulsoup4==4.14.3
# certifi==2026.1.4
# cffi==2.0.0
# charset-normalizer==3.4.4
# curl_cffi==0.13.0
# fastjsonschema==2.21.2
# frozendict==2.4.7
# idna==3.11
# joblib==1.5.3
# jsonschema==4.26.0
# jsonschema-specifications==2025.9.1
# jupyter_core==5.9.1
# lxml==6.0.2
# matplotlib==3.10.8
# mkl-service==2.5.2
# multitasking==0.0.12
# narwhals==2.15.0
# nbformat==5.10.4
# pandas-datareader==0.10.0
# pandas-flavor==0.8.1
# peewee==3.19.0
# platformdirs==4.5.1
# plotly==6.5.2
# protobuf==6.33.4
# pycparser==2.23
# PyQt6==6.9.1
# referencing==0.37.0
# requests==2.32.5
# rpds-py==0.30.0
# scikit-learn==1.8.0
# soupsieve==2.8.3
# threadpoolctl==3.6.0
# traitlets==5.14.3
# urllib3==2.6.3
# websockets==16.0
# xarray==2025.12.0
# yfinance==1.0
